requirements.txt
langchain
langchain_groq
groq
requests
python-dotenv
streamlit